Download the source for doc/filters.texi in each release.
Versions older than 0.6.7 use a different path, but they contain only a couple
of filters, so we can ignore them.

Data is collected from the mirror in GitHub, so we can use the REST API, instead
of cloning the data. It is possible to use the official repository directly with
the git(1) client, but some requests (like downloading a file with git-archive)
are much heavier for the server than the REST API.

Files are stored in `target/storage`, so we can launch the script multiple times
without downloading everything each time.

If the GITHUB_TOKEN environment variable is set, it will be used to authenticate
requests.