Group multiple terms when the definition is shared.
In some tables, a single definition is applied to multiple terms. In such cases,
the definition is associated with the last term of the related sequence.

Now, these definitions are put in the same <dt> element. Thus, we avoid empty
<dd> elements, and the association between terms and definitions is easier to
see.

To detect those cases, we are using <xsl:for-each-group>. Unfortunately, this
feature is not implemented by libxslt1, so we have to move to Saxon.

Saxon is executed as a standalone binary. The overhead of a single execution is
very high (compared to use Nokogiri::XSLT#transform) so we adjust the process to
transform the whole XML in a single invocation, instead of individual sections.

Since Saxon is written in Java, we have to add a JRE to both the development
container and the runner in GitHub Actions.

Finally, there is a bug in Saxon related to emitting empty <a> elements. To
avoid the problem, the anchor now contains the `#` text. This change requires
a different strategy to show the anchors when the pointer is over the section.