Add an icon to the page.
The image was adapted from the [1]logo available in Wikipedia. The pages include
a PNG and a SVG version. The SVG version is the last one, so browsers should
prioritize it.

The PNG version is created from the SVG file, using the rsvg-convert program
from the librsvg2-bin package. Optipng is used to reduce the file size; for this
icon, optipng reduces the size from 3267 to 2546 bytes.

[1]: https://en.wikipedia.org/wiki/FFmpeg#/media/File:FFmpeg_Logo_new.svg