Create a Nix flake to launch the program.
bundix is used to use the gems defined in the Gemfile as Nix dependencies.

The script `.nix/update-gemset.sh` can update the Gemfile.lock and generate a
new `gemset.nix` from bundix.

The script `.nix/sync-ffmpeg-data.sh` is a wrapper to update the snapshot with
data from the FFmpeg repository.

The flake contains all dependencies of the program, so it is possible to invoke
with just `nix run .`.