Keep a copy of FFmpeg data in the repository.
Before this patch, the program needed to download some data from the FFmpeg
repository (like the filters documentation, or the tag's metadata). This method
is incompatible with running in a reproducible environment (like a Nix flake),
and it was also very slow if the data was not already cached.

Now, there is a `--sync-ffmpeg-data` command to create a snapshot of the data,
and store it in a tarball.

The tarball is created with a fixed mtime, so it should generate identical files
if there are no changes in the FFmpeg repository. It is compressed with zstd,
which is faster than xz, though the compression rate slightly worse.

The `target/storage` directory is not used anymore as a cache. Data is kept in
either the path of $FFDOCS_CACHE, or in the `.cache` directory from XDG.